pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/Jonpot/forge"
Issues = "https://github.com/Jonpot/forge/issues"

[tool.setuptools.packages.find]
include = ["backend*", "blocks*", "Forge*"]

[project]
name = "forge"
version = "0.1.0"
description = "Visual block-based experimentation framework for interactive data science pipelines"
authors = [{name = "Jonathan Potter"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
  "pandas>=2.2",
  "pyarrow>=15.0",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "plotly>=5.24",
  "kaleido>=0.2",
  "pydantic>=2.11",
  "fastapi>=0.111",
  "anyio>=4.4",
  "mcp>=1.8",
  "uvicorn[standard]>=0.30",
  "umap-learn>=0.5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.pyright]
executionEnvironments = [
  { root = "tests", extraPaths = ["."], reportAttributeAccessIssue = "none" },
]
